🛠️ Refactor: Improve project structure and enforce centralized error reporting
- Extracted HTML presentation logic into `template.go` (Single Responsibility Principle, Martin).
- Extracted HTTP routing and handling logic from `app.go` to `server.go` (Single Responsibility Principle, Martin).
- Split `ServeHTTP` method into `handleGet` and `handlePost` to reduce Long Method code smell and cognitive complexity (Fowler).
- Introduced centralized error reporter at `internal/reporter/reporter.go` and migrated `main.go` and `app.go` errors, solving retroactive linting rules and `govet` failures.
- Updated `mise.toml` adding `ci` and `test` tasks.
