+++ cmd/ts-fileserver/main.go
@@ -2,10 +2,10 @@ package main
 
 import (
 	"flag"
-	"log"
 
 	"github.com/davecgh/go-spew/spew"
 	fileserver "github.com/lucasew/ts-fileserver"
+	"github.com/lucasew/ts-fileserver/internal/reporter"
 )
 
 func main() {
@@ -22,11 +22,11 @@ func main() {
 
 	app, err := fileserver.NewApp(params)
 	if err != nil {
-		log.Fatalf("failed to initialize application: %w", err)
+		reporter.FatalError("failed to initialize application: %v", err)
 		return
 	}
 	if err := app.Run(); err != nil {
-		log.Fatalf("failed to run app: %w", err)
+		reporter.FatalError("failed to run app: %v", err)
 		return
 	}
 }
